Log an INFO message in `config_once` "happy path"
For now, we're using the log messages from `periodic` as a signal for
whether all is well in `hg-agent`: if there are `ERROR` messages in the
last 10 lines or so, we say "not OK"; otherwise "OK".

An earlier pass removed "unnecessary" logging, with the result that when
everything's good, `periodic` logs are silent. This means we have
nothing to trigger the "OK" state anymore.

This message (re)introduces a "nothing to see here" log in the `config`
task. This is executed every 10s by default, whereas the `heartbeat`
task is executed every 60s.